moving the chromatic font translations into a single file
