--- packages/@react-spectrum/dialog/chromatic/intl/en-US.json
@@ -1,25 +1,0 @@
-{
-  "koji": "Koji",
-  "kojiFoods": "Koji foods",
-  "foodsMakeWithKoji": "Japanese foods, condiments, and beverages made with koji",
-  "kojiColors": "Colors of koji",
-  "white": "white (green)",
-  "yellow": "yellow",
-  "black": "black",
-  "name": "Name",
-  "description": "Description",
-  "soySauce": "Shoyu",
-  "soySauceDescription": "A liquid condiment made from soybeans, koji, and wheat.",
-  "miso": "Miso",
-  "misoDescription": "A traditional Japanese seasoning.",
-  "amazake": "Amazake",
-  "amazakeDescription": "A sweet rice drink.",
-  "sake": "Sake",
-  "sakeDescription": "Rice wine.",
-  "mirin": "Mirin",
-  "mirinDescription": "Sweet rice wine.",
-  "riceWineVinegar": "Rice wine vinegar",
-  "riceWineVinegarDescription": "Vinegar made from rice wine.",
-  "cancel": "Cancel",
-  "confirm": "Confirm"
-}
