--- packages/@react-spectrum/dialog/chromatic/intl/he-IL.json
@@ -1,25 +1,0 @@
-{
-  "koji": "קוג'י",
-  "kojiFoods": "מזונות קוג'י",
-  "foodsMakeWithKoji": "אוכל יפני, תבלינים ומשקאות עשויים עם קוג'י",
-  "kojiColors": "צבעים של קוג'י",
-  "white": "לבן (ירוק)",
-  "yellow": "צהוב",
-  "black": "שָׁחוֹר",
-  "name": "שֵׁם",
-  "description": "תיאור",
-  "soySauce": "שויו",
-  "soySauceDescription": "תבלין נוזלי העשוי מפולי סויה, קוג'י וחיטה.",
-  "miso": "מיסו",
-  "misoDescription": "תיבול יפני מסורתי.",
-  "amazake": "אמזאקה",
-  "amazakeDescription": "משקה אורז מתוק.",
-  "sake": "סאקה",
-  "sakeDescription": "יין אורז.",
-  "mirin": "מירין",
-  "mirinDescription": "יין אורז מתוק.",
-  "riceWineVinegar": "חומץ יין אורז",
-  "riceWineVinegarDescription": "חומץ עשוי מיין אורז.",
-  "cancel": "לְבַטֵל",
-  "confirm": "לְאַשֵׁר"
-}
