--- packages/@react-spectrum/dialog/chromatic/intl/ja-JP.json
@@ -1,25 +1,0 @@
-{
-  "koji": "麹",
-  "kojiFoods": "麹食品",
-  "foodsMakeWithKoji": "麹を使った和食・調味料・飲料",
-  "kojiColors": "麹の色",
-  "white": "白（緑",
-  "yellow": "黄色",
-  "black": "黒",
-  "name": "名前",
-  "description": "説明",
-  "soySauce": "醤油",
-  "soySauceDescription": "大豆、麹、小麦から作られた液体調味料です。",
-  "miso": "味噌",
-  "misoDescription": "日本の伝統的な調味料。",
-  "amazake": "甘酒",
-  "amazakeDescription": "甘いお米のドリンク。",
-  "sake": "酒",
-  "sakeDescription": "お酒。",
-  "mirin": "みりん",
-  "mirinDescription": "味醂。",
-  "riceWineVinegar": "米酢",
-  "riceWineVinegarDescription": "米酒から作られたお酢。",
-  "cancel": "キャンセル",
-  "confirm": "確認"
-}
