--- packages/@react-spectrum/dialog/chromatic/intl/ko-KR.json
@@ -1,25 +1,0 @@
-{
-  "koji": "코지",
-  "kojiFoods": "누룩 식품",
-  "foodsMakeWithKoji": "누룩으로 만든 일본 음식, 조미료 및 음료",
-  "kojiColors": "누룩의 색",
-  "white": "흰색(녹색)",
-  "yellow": "노란색",
-  "black": "검은색",
-  "name": "이름",
-  "description": "설명",
-  "soySauce": "쇼유",
-  "soySauceDescription": "콩, 누룩, 밀로 만든 액체 조미료.",
-  "miso": "된장",
-  "misoDescription": "일본의 전통 조미료.",
-  "amazake": "감주",
-  "amazakeDescription": "달콤한 쌀 음료.",
-  "sake": "동기",
-  "sakeDescription": "막걸리.",
-  "mirin": "미린",
-  "mirinDescription": "달콤한 막걸리.",
-  "riceWineVinegar": "막걸리식초",
-  "riceWineVinegarDescription": "막걸리로 만든 식초.",
-  "cancel": "취소",
-  "confirm": "확인하다"
-}
