--- packages/@react-spectrum/dialog/chromatic/intl/zh-CN.json
@@ -1,25 +1,0 @@
-{
-  "koji": "小司",
-  "kojiFoods": "酒曲食品",
-  "foodsMakeWithKoji": "用曲制成的日本食品、调味品和饮料",
-  "kojiColors": "酒曲的颜色",
-  "white": "白（绿）",
-  "yellow": "黄色的",
-  "black": "黑色的",
-  "name": "姓名",
-  "description": "描述",
-  "soySauce": "酱油",
-  "soySauceDescription": "一种由大豆、曲和小麦制成的液体调味品。",
-  "miso": "味噌",
-  "misoDescription": "日本传统调味料",
-  "amazake": "甘酒",
-  "amazakeDescription": "一种甜米饮料。",
-  "sake": "清酒",
-  "sakeDescription": "米酒。",
-  "mirin": "料酒",
-  "mirinDescription": "甜米酒。",
-  "riceWineVinegar": "米酒醋",
-  "riceWineVinegarDescription": "用米酒制成的醋。",
-  "cancel": "取消",
-  "confirm": "确认"
-}
