--- packages/@react-spectrum/dialog/chromatic/intl/zh-TW.json
@@ -1,25 +1,0 @@
-{
-  "koji": "小司",
-  "kojiFoods": "酒麯食品",
-  "foodsMakeWithKoji": "用曲製成的日本食品、調味品和飲料",
-  "kojiColors": "酒麯的顏色",
-  "white": "白（綠）",
-  "yellow": "黃色的",
-  "black": "黑色的",
-  "name": "姓名",
-  "description": "描述",
-  "soySauce": "醬油",
-  "soySauceDescription": "一種由大豆、曲和小麥製成的液體調味品。",
-  "miso": "味噌",
-  "misoDescription": "日本傳統調味料。",
-  "amazake": "甘酒",
-  "amazakeDescription": "一種甜米飲料。",
-  "sake": "清酒",
-  "sakeDescription": "米酒。",
-  "mirin": "料酒",
-  "mirinDescription": "甜米酒。",
-  "riceWineVinegar": "米酒醋",
-  "riceWineVinegarDescription": "用米酒製成的醋。",
-  "cancel": "取消",
-  "confirm": "確認"
-}
