--- packages/@react-spectrum/form/chromatic/intl/ar-AE.json
@@ -1,29 +1,0 @@
-{
-  "ruminants": "المجترات",
-  "areRuminats": "هذه مجترات.",
-  "whatIsARuminat": "ما هو المجتر؟",
-  "ruminatDefinition": "المجترات هي حيوانات رعي حوافر أو ثدييات قادرة على الحصول على المغذيات من الأطعمة النباتية عن طريق تخميرها في معدة متخصصة قبل الهضم ، بشكل أساسي من خلال الإجراءات الميكروبية.",
-  "cow": "بقرة",
-  "goat": "معزة",
-  "chicken": "فرخة",
-  "sheep": "غنم",
-  "color": "Adobe لون",
-  "colorError": "ليس لون Adobe صالحًا",
-  "farmAnimalCount": "عدد حيوانات المزرعة",
-  "farmCountError": "اختر رقمًا صالحًا",
-  "notRuminatError": "هذا الحيوان ليس مجترًا.",
-  "ageRange": "الفئة العمرية",
-  "cheeseSearch": "البحث عن الجبن",
-  "findCheese": "أدخل جبنًا صالحًا للعثور على الأصل والتاريخ والوصفة.",
-  "farmScore": "نتيجة المزرعة",
-  "amHappy": "هل أنت سعيد؟",
-  "clients": "قائمة العملاء",
-  "clientsError": "أدخل قائمة العملاء.",
-  "clientHeading": "قائمة العملاء",
-  "clientContent": "قائمة مفصولة بفواصل لجميع العملاء.",
-  "farmName": "اسم المزرعة",
-  "brownFenceFarm": "مزرعة السور البني",
-  "farmDescription": "أدخل اسم مزرعتك.",
-  "farmTitle": "الاسم القانوني للمزرعة",
-  "farmLegal": "يجب أن يكون هذا الاسم القانوني للمزرعة وليس اسمًا مستعارًا أو اسم تسويق."
-}
