--- packages/@react-spectrum/form/chromatic/intl/en-US.json
@@ -1,29 +1,0 @@
-{
-  "ruminants": "Ruminants",
-  "areRuminats": "These are ruminants.",
-  "whatIsARuminat": "What is a ruminat?",
-  "ruminatDefinition": "Ruminants are hoofed herbivorous grazing or browsing mammals that are able to acquire nutrients from plant-based food by fermenting it in a specialized stomach prior to digestion, principally through microbial actions.",
-  "cow": "Cow",
-  "goat": "Goat",
-  "chicken": "Chicken",
-  "sheep": "Sheep",
-  "color": "Adobe color",
-  "colorError": "Not a valid Adobe color",
-  "farmAnimalCount": "Farm animal count",
-  "farmCountError": "Choose a valid number",
-  "notRuminatError": "This animal is not a ruminant.",
-  "ageRange": "Age range",
-  "cheeseSearch": "Cheese search",
-  "findCheese": "Enter a valid cheese to find origin, history, and recipe.",
-  "farmScore": "Farm score",
-  "amHappy": "Are you happy?",
-  "clients": "List of clients",
-  "clientsError": "Enter client list.",
-  "clientHeading": "Client list",
-  "clientContent": "Comma seperated list of all clients.",
-  "farmName": "Farm name",
-  "brownFenceFarm": "Brown Fence Farm",
-  "farmDescription": "Enter the name of your farm.",
-  "farmTitle": "Legal name of farm",
-  "farmLegal": "This should be the legal name of the farm and not an alias or marketing name."
-}
