--- packages/@react-spectrum/form/chromatic/intl/he-IL.json
@@ -1,29 +1,0 @@
-{
-  "ruminants": "מעלי גירה",
-  "areRuminats": "אלה בעלי גירה.",
-  "whatIsARuminat": "מה זה גירה?",
-  "ruminatDefinition": "גירה הם יונקים אוכלי עשב פרסות שרועות או גולשים המסוגלים לרכוש חומרים מזינים ממזון צמחי על ידי תסיסתו בקיבה מיוחדת לפני העיכול, בעיקר באמצעות פעולות מיקרוביאליות.",
-  "cow": "פָּרָה",
-  "goat": "עֵז",
-  "chicken": "עוף",
-  "sheep": "כבשים",
-  "color": "Adobe צֶבַע",
-  "colorError": "לא צבע Adobe חוקי",
-  "farmAnimalCount": "ספירת חיות משק",
-  "farmCountError": "בחר מספר חוקי",
-  "notRuminatError": "בעל חיים זה אינו מעלה גירה.",
-  "ageRange": "טווח גילאים",
-  "cheeseSearch": "חיפוש גבינה",
-  "findCheese": "הזן גבינה חוקית כדי למצוא מקור, היסטוריה ומתכון.",
-  "farmScore": "ציון החווה",
-  "amHappy": "האם אתה שמח?",
-  "clients": "רשימת לקוחות",
-  "clientsError": "היכנס לרשימת לקוחות.",
-  "clientHeading": "רשימת לקוחות",
-  "clientContent": "רשימה מופרדת בפסיק של כל הלקוחות.",
-  "farmName": "שם החווה",
-  "brownFenceFarm": "חוות גדר חומה",
-  "farmDescription": "הזן את שם החווה שלך.",
-  "farmTitle": "השם החוקי של החווה",
-  "farmLegal": "זה צריך להיות השם החוקי של החווה ולא כינוי או שם שיווקי."
-}
