--- packages/@react-spectrum/form/chromatic/intl/ja-JP.json
@@ -1,29 +1,0 @@
-{
-  "ruminants": "反芻動物",
-  "areRuminats": "これらは反芻動物です。",
-  "whatIsARuminat": "反芻動物とは？",
-  "ruminatDefinition": "反芻動物は、主に微生物の作用を通じて、消化前に特殊な胃で発酵させることにより、植物ベースの食物から栄養素を取得できる草食性の草食または閲覧する有蹄動物です。",
-  "cow": "牛",
-  "goat": "ヤギ",
-  "chicken": "鶏",
-  "sheep": "羊",
-  "color": "Adobe 色",
-  "colorError": "有効な Adob​​e カラーではありません",
-  "farmAnimalCount": "家畜数",
-  "farmCountError": "有効な番号を選択してください",
-  "notRuminatError": "この動物は反芻動物ではありません",
-  "ageRange": "年齢層",
-  "cheeseSearch": "チーズ検索",
-  "findCheese": "有効なチーズを入力して、起源、歴史、レシピを検索してください。",
-  "farmScore": "農場スコア",
-  "amHappy": "幸せですか？",
-  "clients": "クライアント一覧",
-  "clientsError": "顧客リストを入力します。",
-  "clientHeading": "クライアントリスト",
-  "clientContent": "すべてのクライアントのカンマ区切りリスト。",
-  "farmName": "農場名",
-  "brownFenceFarm": "ブラウン フェンス ファーム",
-  "farmDescription": "農場の名前を入力します。",
-  "farmTitle": "農場の正式名称",
-  "farmLegal": "これは農場の正式な名前であり、エイリアスやマーケティング名ではありません。"
-}
