--- packages/@react-spectrum/form/chromatic/intl/ko-KR.json
@@ -1,29 +1,0 @@
-{
-  "ruminants": "반추 동물",
-  "areRuminats": "이들은 반추 동물입니다.",
-  "whatIsARuminat": "반추동물이란?",
-  "ruminatDefinition": "반추동물은 발굽이 있는 초식성 방목 또는 브라우징 포유류로, 주로 미생물 작용을 통해 소화하기 전에 특수 위장에서 식물성 식품을 발효시켜 영양분을 얻을 수 있습니다.",
-  "cow": "암소",
-  "goat": "염소",
-  "chicken": "닭",
-  "sheep": "양",
-  "color": "Adobe 색상",
-  "colorError": "유효한 Adobe 색상이 아닙니다.",
-  "farmAnimalCount": "농장 동물 수",
-  "farmCountError": "유효한 번호를 선택하세요",
-  "notRuminatError": "이 동물은 반추 동물이 아닙니다.",
-  "ageRange": "연령대",
-  "cheeseSearch": "치즈 검색",
-  "findCheese": "원산지, 역사 및 조리법을 찾으려면 유효한 치즈를 입력하십시오.",
-  "farmScore": "농장 점수",
-  "amHappy": "행복하니?",
-  "clients": "클라이언트 목록",
-  "clientsError": "클라이언트 목록을 입력하십시오.",
-  "clientHeading": "클라이언트 목록",
-  "clientContent": "모든 클라이언트의 쉼표로 구분된 목록입니다.",
-  "farmName": "농장 이름",
-  "brownFenceFarm": "갈색 울타리 농장",
-  "farmDescription": "농장 이름을 입력하세요.",
-  "farmTitle": "농장의 법적 이름",
-  "farmLegal": "별칭이나 마케팅 이름이 아니라 팜의 법적 이름이어야 합니다."
-}
