--- packages/@react-spectrum/form/chromatic/intl/zh-CN.json
@@ -1,29 +1,0 @@
-{
-  "ruminants": "反刍动物",
-  "areRuminats": "这些是反刍动物。",
-  "whatIsARuminat": "什么是反刍动物？",
-  "ruminatDefinition": "反刍动物是有蹄的草食性食草或浏览哺乳动物，它们能够通过在消化前在专门的胃中发酵来从植物性食物中获取营养，主要是通过微生物作用。",
-  "cow": "奶牛",
-  "goat": "山羊",
-  "chicken": "鸡",
-  "sheep": "羊",
-  "color": "Adobe 颜色",
-  "colorError": "不是有效的 Adob​​e 颜色",
-  "farmAnimalCount": "农场动物数量",
-  "farmCountError": "选择一个有效的号码",
-  "notRuminatError": "这种动物不是反刍动物。",
-  "ageRange": "年龄范围",
-  "cheeseSearch": "奶酪搜索",
-  "findCheese": "输入有效的奶酪以查找产地、历史和配方。",
-  "farmScore": "农场分数",
-  "amHappy": "你快乐吗？",
-  "clients": "客户名单",
-  "clientsError": "输入客户名单。",
-  "clientHeading": "客户名单",
-  "clientContent": "所有客户端的逗号分隔列表。",
-  "farmName": "农场名称",
-  "brownFenceFarm": "布朗栅栏农场",
-  "farmDescription": "输入您农场的名称。",
-  "farmTitle": "农场的法定名称",
-  "farmLegal": "这应该是农场的法定名称，而不是别名或营销名称。"
-}
