--- packages/@react-spectrum/form/chromatic/intl/zh-TW.json
@@ -1,29 +1,0 @@
-{
-  "ruminants": "反芻動物",
-  "areRuminats": "這些是反芻動物。",
-  "whatIsARuminat": "什麼是反芻動物？",
-  "ruminatDefinition": "反芻動物是有蹄的草食性食草或瀏覽哺乳動物，它們能夠通過在消化前在專門的胃中發酵來從植物性食物中獲取營養，主要是通過微生物作用。",
-  "cow": "奶牛",
-  "goat": "山羊",
-  "chicken": "雞",
-  "sheep": "羊",
-  "color": "Adobe 顏色",
-  "colorError": "不是有效的 Adob​​e 顏色",
-  "farmAnimalCount": "農場動物數量",
-  "farmCountError": "選擇一個有效的號碼",
-  "notRuminatError": "這種動物不是反芻動物。",
-  "ageRange": "年齡範圍",
-  "cheeseSearch": "奶酪搜索",
-  "findCheese": "輸入有效的奶酪以查找產地、歷史和配方。",
-  "farmScore": "農場分數",
-  "amHappy": "你快樂嗎？",
-  "clients": "客戶名單",
-  "clientsError": "輸入客戶名單。",
-  "clientHeading": "客戶名單",
-  "clientContent": "所有客戶端的逗號分隔列表。",
-  "farmName": "農場名稱",
-  "brownFenceFarm": "布朗柵欄農場",
-  "farmDescription": "輸入您農場的名稱。",
-  "farmTitle": "農場的法定名稱",
-  "farmLegal": "這應該是農場的法定名稱，而不是別名或營銷名稱。"
-}
