--- packages/@react-spectrum/menu/chromatic/intl/ar-AE.json
@@ -1,6 +1,0 @@
-{
-  "button": "يحرر",
-  "copy": "ينسخ",
-  "cut": "يقطع",
-  "paste": "معجون"
-}
