--- packages/@react-spectrum/menu/chromatic/intl/en-US.json
@@ -1,6 +1,0 @@
-{
-  "button": "Edit",
-  "copy": "Copy",
-  "cut": "Cut",
-  "paste": "Paste"
-}
