--- packages/@react-spectrum/menu/chromatic/intl/he-IL.json
@@ -1,6 +1,0 @@
-{
-  "button": "לַעֲרוֹך",
-  "copy": "עותק",
-  "cut": "גזירה",
-  "paste": "לְהַדבִּיק"
-}
