--- packages/@react-spectrum/menu/chromatic/intl/ja-JP.json
@@ -1,6 +1,0 @@
-{
-  "button": "編集",
-  "copy": "コピー",
-  "cut": "切る",
-  "paste": "ペースト"
-}
