--- packages/@react-spectrum/menu/chromatic/intl/ko-KR.json
@@ -1,6 +1,0 @@
-{
-  "button": "편집하다",
-  "copy": "복사",
-  "cut": "자르다",
-  "paste": "반죽"
-}
