--- packages/@react-spectrum/menu/chromatic/intl/zh-CN.json
@@ -1,6 +1,0 @@
-{
-  "button": "编辑",
-  "copy": "复制",
-  "cut": "切",
-  "paste": "粘贴"
-}
