--- packages/@react-spectrum/menu/chromatic/intl/zh-TW.json
@@ -1,6 +1,0 @@
-{
-  "button": "編輯",
-  "copy": "複製",
-  "cut": "切",
-  "paste": "粘貼"
-}
